tests/wpt/tests/css/css-view-transitions/parsing/view-transition-class-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS View Transitions Test: view-transition-class with valid values</title>
<link rel="author" title="Vladimir Levin" href="mailto:vmpstr@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-view-transitions-1/">
<meta name="assert" content="view-transition-class supports custom tags">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("view-transition-class", "none");
test_valid_value("view-transition-class", "foo");
test_valid_value("view-transition-class", "bar");
test_valid_value("view-transition-class", "baz");
test_valid_value("view-transition-class", "foo bar");
test_valid_value("view-transition-class", "foo bar baz");
</script>
</body>
</html>
